pyproject.toml
[tool.poetry]
name = "my-fastapi-project"
version = "0.1.0"
description = "FastAPI project"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
grpcio = "^1.69.0"
pydantic = "^2.10.4"
uvicorn = "^0.34.0"
google-generativeai = "^0.8.3"
